migrate `PageTitleTabs`
