+++ packages/app/src/kits/Structure/PageTitle/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { PageTitleTabsProps } from '../PageTitleTabs/types';
+import type { PageTitleTabsProps } from 'ui-structure/src/types';
 
 export type PageTitleProps = PageTitleTabsProps & {
   colorSecondary?: boolean;
