+++ packages/app/src/pages/Nominate/index.tsx
@@ -6,7 +6,6 @@ import { NominationGeo } from './NominationGeo';
 import { Wrapper } from './Wrappers';
 import { PageTitle } from '../../kits/Structure/PageTitle';
 import { useTranslation } from 'react-i18next';
-import type { PageTitleTabProps } from 'kits/Structure/PageTitleTabs/types';
 import { useEffect, useState } from 'react';
 import { usePlugins } from 'contexts/Plugins';
 
@@ -15,23 +14,6 @@ export const Nominate = () => {
   const { pluginEnabled } = usePlugins();
 
   const [activeTab, setActiveTab] = useState<number>(0);
-
-  // Only include tabs if the Polkawatch plugin is enabled.
-  let tabs: PageTitleTabProps[] | undefined = undefined;
-  if (pluginEnabled('polkawatch')) {
-    tabs = [
-      {
-        title: t('overview', { ns: 'base' }),
-        active: activeTab === 0,
-        onClick: () => setActiveTab(0),
-      },
-      {
-        title: t('decentralization', { ns: 'base' }),
-        active: activeTab === 1,
-        onClick: () => setActiveTab(1),
-      },
-    ];
-  }
 
   // Go back to the first tab if the Polkawatch plugin is disabled.
   useEffect(() => {
@@ -42,7 +24,25 @@ export const Nominate = () => {
 
   return (
     <Wrapper>
-      <PageTitle title={t('nominate.nominate', { ns: 'pages' })} tabs={tabs} />
+      <PageTitle
+        title={t('nominate.nominate', { ns: 'pages' })}
+        tabs={
+          pluginEnabled('polkawatch')
+            ? [
+                {
+                  title: t('overview', { ns: 'base' }),
+                  active: activeTab === 0,
+                  onClick: () => setActiveTab(0),
+                },
+                {
+                  title: t('decentralization', { ns: 'base' }),
+                  active: activeTab === 1,
+                  onClick: () => setActiveTab(1),
+                },
+              ]
+            : undefined
+        }
+      />
       {activeTab == 0 && <Active />}
       {activeTab == 1 && <NominationGeo />}
     </Wrapper>
