+++ packages/ui-structure/src/PageTitleTabs/index.module.scss
@@ -1,0 +1,35 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.pageTitleTabs {
+  border-bottom: 1px solid var(--border-primary-color);
+  overflow: hidden;
+  height: 3.6rem;
+  margin-top: 0.9rem;
+  max-width: 100%;
+
+  // NOTE: This is the same as `PageWidthMediumThreshold` constant in `consts` package. Could
+  // migrate to scss variables in the future.
+  @media (max-width: 1150px) {
+    margin-top: 0.5rem;
+  }
+}
+
+.pageTitleTabs_inline {
+  border-bottom: none;
+}
+
+.pageTitleTabs_sticky {
+  border-bottom: 0;
+  margin-top: 0.5rem;
+}
+
+.pageTitleTabs_scroll {
+  width: 100%;
+  height: 4.5rem;
+  overflow: auto hidden;
+}
+
+.pageTitleTabs_inner {
+  display: flex;  
+}
