+++ packages/ui-structure/src/PageTitleTabs/index.tsx
@@ -1,0 +1,50 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+import { ButtonTab } from 'ui-buttons';
+import type { PageTitleProps, PageTitleTabProps } from '../types';
+import classNames from 'classnames';
+import classes from './index.module.scss';
+
+/**
+ * @name PageTitleTabs
+ * @summary The element in a page title, inculding ButtonTabs.
+ */
+export const PageTitleTabs = ({
+  sticky,
+  tabs = [],
+  inline = false,
+  tabClassName,
+  colorSecondary,
+}: PageTitleProps) => {
+  const buttonClasses = classNames(classes.pageTitleTabs, {
+    [classes.pageTitleTabs_inline]: inline,
+    [classes.pageTitleTabs_sticky]: sticky,
+  });
+
+  return (
+    <section className={buttonClasses}>
+      <div className={classes.pageTitleTabs_scroll}>
+        <div className={classes.pageTitleTabs_inner}>
+          {tabs.map(
+            (
+              { active, onClick, title, badge, disabled }: PageTitleTabProps,
+              i: number
+            ) => (
+              <ButtonTab
+                className={tabClassName}
+                active={!!active}
+                key={`page_tab_${i}`}
+                onClick={() => onClick()}
+                title={title}
+                colorSecondary={colorSecondary}
+                badge={badge}
+                disabled={disabled === undefined ? false : disabled}
+              />
+            )
+          )}
+        </div>
+      </div>
+    </section>
+  );
+};
