+++ packages/ui-structure/src/index.tsx
@@ -8,3 +8,4 @@ export * from './Main';
 export * from './Page';
 export * from './PageRow';
 export * from './PageHeading';
+export * from './PageTitleTabs';
